Factor the diff override command
This is the command that both performs a diff of a single override as
well as offers the interactive part of bulk diffing an entire package
by allowing the user to select the package to diff.

This pulls the command out into its own file, ever shortening the
override_audit.py file.

In order to stop an imminent collision with a command that is going
to be added when the context command for override options becomes a
series of commands instead of just a single one overloaded with an
action property, along the way the command has changed names from
override_audit_diff_override to override_audit_diff_single.

The operation of the command otherwise remains the same.
