+++ Default.sublime-commands
@@ -67,11 +67,11 @@
     },
     {
         "caption": "OverrideAudit: Diff Single Override",
-        "command": "override_audit_diff_override"
+        "command": "override_audit_diff_single"
     },
     {
         "caption": "OverrideAudit: Bulk Diff Single Package",
-        "command": "override_audit_diff_override",
+        "command": "override_audit_diff_single",
         "args": { "bulk": true }
     },
     {
