+++ Main.sublime-menu
@@ -10,8 +10,8 @@
                     { "caption": "Package Report", "command": "override_audit_package_report"},
                     { "caption": "Override Report", "command": "override_audit_override_report"},
                     { "caption": "Override Report (Only expired)", "command": "override_audit_override_report", "args": {"only_expired": true}},
-                    { "caption": "Diff Single Override...", "command": "override_audit_diff_override"},
-                    { "caption": "Bulk Diff: Single Package...", "command": "override_audit_diff_override", "args": {"bulk": true}},
+                    { "caption": "Diff Single Override...", "command": "override_audit_diff_single"},
+                    { "caption": "Bulk Diff: Single Package...", "command": "override_audit_diff_single", "args": {"bulk": true}},
                     { "caption": "Bulk Diff: All Packages", "command": "override_audit_diff_package"},
                     { "caption": "Refresh Current Report", "command": "override_audit_context_report"},
                     { "caption": "-" }
