+++ src/commands/__init__.py
@@ -1,11 +1,12 @@
 from ...bootstrap import reload
 
 reload("src.commands", ["package_report", "override_report", "diff_package",
-       "modify_mark"])
+       "diff_single", "modify_mark"])
 
 from .package_report import OverrideAuditPackageReportCommand
 from .override_report import OverrideAuditOverrideReportCommand
 from .diff_package import OverrideAuditDiffPackageCommand
+from .diff_single import OverrideAuditDiffSingleCommand
 from .modify_mark import OverrideAuditModifyMarkCommand
 
 __all__ = [
@@ -13,6 +14,7 @@ __all__ = [
     "OverrideAuditPackageReportCommand",
     "OverrideAuditOverrideReportCommand",
     "OverrideAuditDiffPackageCommand",
+    "OverrideAuditDiffSingleCommand",
 
     # General
     "OverrideAuditModifyMarkCommand"
